bigger 3v3 tracks
not sure if LED supply is better or worse now - it's away from the opamp (that's good), but now might cause interference on the temp. sensors (that's bad). 

If LED isn't switched much, and/or if temperature accuracy isn't critical, this layout should be OK.

Might still be worth moving some of the 3v3 capacitors over to the phototransistors though.